All windows connected (may be some problems)

+++ KM/KM/Model/Input.cs
@@ -62,7 +62,7 @@ namespace KM
     public class TableObject
     {
         public int Number { get; set; }
-        public double[] X { get; set; }
+        public int[] X { get; set; }
         public double[] Y { get; set; }
     }
